requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.5.0
Django==1.11.7
djangorestframework==3.8.2
get==0.0.21
gunicorn==19.8.1
idna==2.6
numpy==1.14.2
pandas==0.22.0
post==0.0.13
psycopg2==2.7.4
public==0.0.38
python-dateutil==2.7.2
pytz==2017.3
query-string==0.0.12
request==0.0.13
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.1
setupfiles==0.0.50
six==1.11.0
sklearn==0.0
urllib3==1.22
urllib5==5.0.0
virtualenv==15.1.0
virtualenvwrapper-win==1.2.3
whitenoise==3.3.1
